+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 #### Anomalies
 
 * Fixes the warning for arm64_32 architecture (watchOS 5) in RxCocoa for Swift versions priror to 4.2.
+* Fixes issues with CocoaPods and Swift 4.2 version.
 
 ## [4.3.0](https://github.com/ReactiveX/RxSwift/releases/tag/4.3.0)